+++ invokeai/frontend/web/src/app/store/nanostores/headerComponent.ts
@@ -1,0 +1,4 @@
+import { atom } from 'nanostores';
+import { ReactNode } from 'react';
+
+export const $headerComponent = atom<ReactNode | undefined>(undefined);
