+++ invokeai/frontend/web/src/app/store/nanostores/index.ts
@@ -1,0 +1,3 @@
+/**
+ * For non-serializable data that needs to be available throughout the app, or when redux is not appropriate, use nanostores.
+ */
